Added test for baseline seq-to-seq training.
